Refactor list pools
