+++ Packages/com.deltation.di-framework/Assets/DELTation/DIFramework/Runtime/Pooling/ListPool.cs
@@ -1,16 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using JetBrains.Annotations;
-using UnityEngine;
 
-namespace DELTation.DIFramework.Resolution
+namespace DELTation.DIFramework.Pooling
 {
-    internal static class ComponentListPool
+    internal static class ListPool<T>
     {
-        public static List<MonoBehaviour> Rent()
+        public static List<T> Rent()
         {
             if (FreeLists.Count == 0)
-                return new List<MonoBehaviour>();
+                return new List<T>();
 
             var lastListIndex = FreeLists.Count - 1;
             var lastList = FreeLists[lastListIndex];
@@ -20,7 +19,7 @@ namespace DELTation.DIFramework.Resolution
             return lastList;
         }
 
-        public static void Return([NotNull] List<MonoBehaviour> list)
+        public static void Return([NotNull] List<T> list)
         {
             if (list == null) throw new ArgumentNullException(nameof(list));
 
@@ -31,7 +30,7 @@ namespace DELTation.DIFramework.Resolution
             FreeListsSet.Add(list);
         }
 
-        private static readonly HashSet<List<MonoBehaviour>> FreeListsSet = new HashSet<List<MonoBehaviour>>();
-        private static readonly List<List<MonoBehaviour>> FreeLists = new List<List<MonoBehaviour>>();
+        private static readonly HashSet<List<T>> FreeListsSet = new HashSet<List<T>>();
+        private static readonly List<List<T>> FreeLists = new List<List<T>>();
     }
 }
