--- Packages/com.deltation.di-framework/Assets/DELTation/DIFramework/Runtime/Resolution/ComponentListPool.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a7a50e66d4d846c09132050461d2ff7f
-timeCreated: 1615919098
